Automatically create container for adapter, and get API routes for jsonld-storage.js
