+++ docker-compose.yaml
@@ -40,7 +40,8 @@ services:
     container_name: middleware
     restart: always
     depends_on:
-      - "fuseki"
+      - fuseki
+      - mongo
     volumes:
       - ./src/middleware:/main/
     networks:
@@ -60,7 +61,7 @@ services:
     container_name: frontend
     restart: always
     depends_on:
-      - "middleware"
+      - middleware
     environment:
       - PORT=5000
     volumes:
